services/o2c-api/requirements.txt
Flask==3.0.3
pika==1.3.2
PyJWT==2.9.0
prometheus-client==0.20.0
